+++ YooniK.Face/YooniK.Face.Client.Sample/Program.cs
@@ -35,19 +35,19 @@ namespace YooniK.Face.Sample
                 string secondPhotoInBase64 = ImageToBase64String(secondPhoto);
 
                 // Verifies the faces similarity between two images in base 64
-                VerifyImagesResponse verifyImages = await faceClient.VerifyImagesAsync(firstPhotoInBase64, secondPhotoInBase64);
-                Console.WriteLine($"Similarity Score: { verifyImages.Score }");
+                MatchingResponse imagesMatchingResponse = await faceClient.VerifyImagesAsync(firstPhotoInBase64, secondPhotoInBase64);
+                Console.WriteLine($"Similarity Score: { imagesMatchingResponse.Score }");
 
                 // Processes all the image containing faces, and returns its information in a list. This photos only contains one face. 
-                List<ProcessResponse> process = await faceClient.ProcessAync(firstPhotoInBase64);
+                List<ProcessResponse> process = await faceClient.ProcessAsync(firstPhotoInBase64);
                 string firstPhotoTemplate = process.Count == 1 ? process[0].Template : null;
 
-                List<ProcessResponse> process2 = await faceClient.ProcessAync(secondPhotoInBase64);
+                List<ProcessResponse> process2 = await faceClient.ProcessAsync(secondPhotoInBase64);
                 string secondPhotoTemplate = process2.Count == 1 ? process2[0].Template : null;
 
                 // Verifies the faces similarity between the extracted biometric template from the processed images
-                VerifyResponse verify = await faceClient.VerifyAsync(firstPhotoTemplate, secondPhotoTemplate);
-                Console.WriteLine($"Similarity Score (w/Template): {verify.Score}");
+                MatchingResponse templateMatchingResponse = await faceClient.VerifyAsync(firstPhotoTemplate, secondPhotoTemplate);
+                Console.WriteLine($"Similarity Score (w/Template): {templateMatchingResponse.Score}");
 
                 // Unique identifiers for the Gallery
                 string galleryGuid = Guid.NewGuid().ToString();