+++ YooniK.Face/YooniK.Face.Client/Client/FaceException.cs
@@ -1,0 +1,13 @@
+﻿using System;
+
+namespace YooniK.Face.Client.Client.FaceException
+{
+
+    [Serializable]
+    class FaceException : Exception
+    {
+        public FaceException(string message) : base(message)
+        {
+        }
+    }
+}