+++ YooniK.Face/YooniK.Face.Client/Models/Responses/Face/VerifyIdResponse.cs
@@ -1,7 +1,0 @@
-﻿
-namespace YooniK.Face.Client.Models.Responses
-{
-    public class VerifyIdResponse : MatchingResponse
-    {
-    }
-}